requirements.txt
langchain==0.1.16
langchain-community
langchain-core
langchain-groq
groq

chromadb

pypdf

tavily-python

streamlit
python-dotenv